pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embodied-robot-rag"
version = "0.1.0"
description = "Embodied robot knowledge QA system built with FastAPI, LangChain RAG, prompt engineering, and DeepSeek."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.12,<1.0.0",
  "langchain>=0.3.23,<1.0.0",
  "langchain-chroma>=0.2.6,<1.0.0",
  "langchain-core>=0.3.52,<1.0.0",
  "langchain-deepseek>=0.1.3,<1.0.0",
  "langchain-text-splitters>=0.3.8,<1.0.0",
  "numpy>=2.2.4,<3.0.0",
  "pydantic-settings>=2.8.1,<3.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.3.5,<9.0.0",
  "ruff>=0.11.2,<1.0.0",
]
hf = [
  "langchain-huggingface>=0.1.2,<1.0.0",
  "sentence-transformers>=3.4.1,<4.0.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"app/api/routes/*.py" = ["B008"]
